Fix jtest compile, quickly

+++ tools/jtest/jtest.cc
@@ -3801,7 +3801,7 @@ UrlHashTable::UrlHashTable() {
     // size as the size
     if (!len) 
       panic("zero size URL Hash Table\n");
-    if (len != URL_HASH_BYTES) {
+    if (len != (unsigned long)URL_HASH_BYTES) {
       fprintf(stderr, 
               "FATAL: hash file length (%lu) != URL_HASH_BYTES (%lu)\n",
               len, (unsigned long)URL_HASH_BYTES);
